+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed a congestion bug during MAVLink log downloads by not requesting more
+  than 512 `LOG_DATA` chunks at a time.
+
 ## [2.17.0] - 2023-10-29
 
 ### Added